feat: add release notes (#350)
This includes the release notes in the reference manual, see
https://pr-350--lean-reference-manual-review.netlify.app/releases/v4.17.0/
for a preview.

It uses Verso’s markdown parser so that one does not have to manually
edit the text for verso-compatibility (our commit messages are messy).

It uses one file and page per release, for easy linking and that it
doesn’t grow too large, and (in particular) that you can scroll or
Ctrl-F without losing track of which version this relates to.

Co-authored-by: David Thrane Christiansen <david@lean-fro.org>