chore: style consistency and faster builds in grind chapter (#554)
* Adds missing annotations to text in the grind chapter, enabling
hyperlinks and related features

* Splits the chapter into files, which allows better build times due to
incrementality and parallelism

* Uses a different mechanism for including the longer code examples.
It's now more convenient to work with, but not quite as good as it
should be.

* Error messages are truncated where we want them to be (e.g. omission
of the table on a WF recursion error)

* grind errors include collapsible trace views. The sections open by
default are deliberately configured.

* It's possible to cite machine-checked sub-sections of the grind
messages, such as showing only the cutsat part

* The style was generally made more consistent with the rest of the
manual

* In some sections, content was reordered to group implementation
details together and present the user-facing features first


@kim-em I believe this addresses the majority of the Verso features we
talked about. Do you want to give it a read-through before merging, or
should I just go ahead?