feat: live links for uses of example directive (#609)
Add some important pieces of the implementation of links to
http://live.lean-lang.org/ without actually enabling it yet. This PR
should be a no-op on the main content of the reference manual, although
it has the side effect of producing an additional directory
`extracted-examples/` inside the output directory during HTML rendering
of the reference manual.

This includes:
- Helper functions to traverse verso content finding code blocks that
are not `-keep` or `+error`, suitably concatenating their contents and
cleaning up whitespace.
 - Filtering out `:::example`s that have no such blocks to display.
- LZ compression compatible with `http://live.lean-lang.org/#codez=`
urls
- Extracting all examples to files in `_out` during document rendering
so that we can run CI on them
 - CSS for displaying the link when we want to.
 - One incidental fix of a scope bug in documentation in `Monads.lean`
 
 Progress towards https://github.com/leanprover/verso/issues/569 .