refactor: import the Lean reference manual as a dependency instead of a fork
I wanted to update the manual to a newer Lean version, but this proved quite annoying since we have to fix a lot of merge conflicts between the files we copied and those we replaced. I propose having the Lean manual as a dependency instead of forking from there.

All our own files in Manual are renamed to MathlibManual to avoid naming conflicts. Also rename the package from verso-manual to mathlib-manual.

(David would be happy to extract some parts of the manual as a standalone library if we ask, I expect.)