feat: walk-through of type class resolution (#663)
Gives an example for the `#synth` interaction, but also gives a bit of a
walk-through of the existing type class resolution tracing mechanism
inspired by the discussion from Gray @willcrichton and Krishnamurthi's
[PLDI paper](https://maint.dlnext.acm.org/).

I think the actual tracing view could be improved with some observations
from their work, but a bit of explanatory text will go some distance.

The two additions are an example of using `#synth` in [the overview of
interacting with
lean](https://pr-663--lean-reference-manual-review.netlify.app/Interacting-with-Lean/#hash-synth)
and a walk-through of synth instance tracing in [Instance Search
Problems](https://pr-663--lean-reference-manual-review.netlify.app/Type-Classes/Instance-Synthesis/#instance-search).

---------

Co-authored-by: David Thrane Christiansen <david@lean-fro.org>