fix: function namespace correction, extended example for generalized field notation (#679)
I noticed the docs were inconsistent with the implementation on the
topic of the correctness of `Nat.add.uncurry` and other
`Function`-namespace definitions, so I fixed that and also added an
example describing the working direction of type unfolding. The
example's getting a little log here but I think it's still reasonable.

I also changed `root` to `admin` as the username example, since `root`
has other meanings for namespacing in particular and I got thrown off
when I read it at first.